Add error checking to bindCommands()
If you try to bind a command and it fails, that command will not be
added, but it will not disrupt everything. (An error will be shown in
the console.)

Also changed the default command to be a default KEYBOARD command, and
made getEventType() slightly more efficient.

Addresses error 2: https://github.com/EvanHahn/Fiesta.js/issues/2