Move logic of graphics drawings to the graphics, not the playground
The playground used to hold logic for drawing things, and it only knew
how to draw sprites. The logic for drawing things on a canvas has now
been moved into Fiesta.Graphic (and its subclasses). This means that
the playground will do nothing differently between 2D and 3D; the only
difference should be in the graphics' implementation.

Also in this commit:

- Shortened a comment about the JS.Class alias
- Improved header of the Browser section