Basic collisions (buggy!)
Every object has a rectangular (technically it's a rectangular prism)
bounding box (technically it's a bounding rectangular prism). When two
objects' bounding boxes collide, the momentum of one object is
transferred into the other.

Pretty basic collision detection for now. Luckily, it's written in
such a way that it shouldn't require major rewrites to change the
algorithm. I hope.

It's buggy; this should be worked on.

Also in this commit:

- Playground.frame is cleaned up a little
- Ball demo gives the box a mass of 5 to illustrate momentum

Addresses issue 20, but not completely.
https://github.com/EvanHahn/Fiesta.js/issues/20