Add dormant PPI API
Instead of setting the speed to 500 pixels/s (which varies per display),
you should be setting the speed to 1 inch/sec. This begins to address
that by adding getPPI(). It gets the pixels per inch.

As far as I can imagine, this is an internal method, but it's possible
that it could be used by other people.

Another method is the calculatePPI method, which recalculates the PPI.
At the moment, this is only called once, and then its value is
cached.

This code is a cleaned-up version of the code at the link below:
http://snipplr.com/view/20912/
I couldn't find the licensing, so I assumed it was free license.

Begins to address issue 10:
http://github.com/EvanHahn/Fiesta.js/issues/10