No more separation of entities
In the past, there was an Entity, a Locatable Entity, and a Physical
Entity. I looked at this for awhile, and concluded that this was a
mistake -- too much inheritance was going on. No longer!

All of my unit tests pass so far, but that doesn't mean anything. I'm
pretty sure the demos still break, and because this is a fairly large
change, I'd believe other things break as well.

Still, this change is a good idea. It makes things much easier to
understand.

+ New: Entities now have a solidity property, accessible by isSolid()
       and setSolidity(true).
+ New: Config for whether entities default to solid or not
* Changed: Physical and Locatable entities are now lumped into Entity